Integrated application date for balance and only allowed journal voucher for allowed domains
