+++ apps/web-giddh/src/app/all-modules/all-modules.component.html
@@ -6,7 +6,7 @@
                 <p><a href="javascript:void(0)" [routerLink]="['/pages/contact/customer']" [queryParams]="{tab: 'customer', tabIndex: 0}">Customer</a></p>
                 <p><a href="javascript:void(0)" [routerLink]="['/pages/contact/aging-report']">Aging Report </a></p>
                 <p><a href="javascript:void(0)" [routerLink]="['/pages/contact/vendor']" [queryParams]="{tab: 'vendor', tabIndex: 0}">Vendor</a></p>
-                <p><a href="javascript:void(0)" [routerLink]="['/pages/journal-voucher']">Journal Voucher *</a></p>
+                <p><a href="javascript:void(0)" *ngIf="isAllowedForBetaTesting" [routerLink]="['/pages/journal-voucher']">Journal Voucher *</a></p>
                 <p><a href="javascript:void(0)" (click)="showManageGroupsModal();">Create Account</a></p>
             </div>
             <div class="col-md-3 col-sm-3 mb-5 col-xs-6">
